fix(angular): use hostReplacementPaths function instead of host
The new approach fixes the problems with watching platform-specific
files.

depends on #611
fixes #601
