+++ demo/AngularApp/webpack.config.js
@@ -3,7 +3,6 @@ const { join, relative, resolve, sep } = require("path");
 const webpack = require("webpack");
 const nsWebpack = require("nativescript-dev-webpack");
 const nativescriptTarget = require("nativescript-dev-webpack/nativescript-target");
-const { PlatformReplacementHost } = require("nativescript-dev-webpack/host/platform");
 const CleanWebpackPlugin = require("clean-webpack-plugin");
 const CopyWebpackPlugin = require("copy-webpack-plugin");
 const { BundleAnalyzerPlugin } = require("webpack-bundle-analyzer");
@@ -227,7 +226,7 @@ module.exports = env => {
             new NativeScriptWorkerPlugin(),
 
             new AngularCompilerPlugin({
-                host: platformHost,
+                hostReplacementPaths: nsWebpack.getResolver([platform, "tns"]),
                 entryModule: resolve(appPath, "app.module#AppModule"),
                 tsConfigPath: join(__dirname, "tsconfig.esm.json"),
                 skipCodeGeneration: !aot,
