+++ package.json
@@ -92,11 +92,8 @@
     "sass-loader": "~7.0.1"
   },
   "devDependencies": {
-    "@angular-devkit/core": "^0.7.0-rc.0",
     "@types/node": "^8.0.0",
     "conventional-changelog-cli": "^1.3.22",
-    "rxjs": "^6.2.0",
-    "source-map-support": "^0.5.0",
     "typescript": "~2.7.2"
   }
 }
