+++ templates/webpack.angular.js
@@ -3,7 +3,6 @@ const { join, relative, resolve, sep } = require("path");
 const webpack = require("webpack");
 const nsWebpack = require("nativescript-dev-webpack");
 const nativescriptTarget = require("nativescript-dev-webpack/nativescript-target");
-const { PlatformReplacementHost } = require("nativescript-dev-webpack/host/platform");
 const CleanWebpackPlugin = require("clean-webpack-plugin");
 const CopyWebpackPlugin = require("copy-webpack-plugin");
 const { BundleAnalyzerPlugin } = require("webpack-bundle-analyzer");
@@ -22,9 +21,6 @@ module.exports = env => {
     if (!platform) {
         throw new Error("You need to provide a target platform!");
     }
-
-    const extensions = ["tns", platform];
-    const platformHost = new PlatformReplacementHost(extensions);
 
     const projectRoot = __dirname;
 
@@ -226,7 +222,7 @@ module.exports = env => {
             new NativeScriptWorkerPlugin(),
 
             new AngularCompilerPlugin({
-                host: platformHost,
+                hostReplacementPaths: nsWebpack.getResolver([platform, "tns"]),
                 entryModule: resolve(appPath, "app.module#AppModule"),
                 tsConfigPath: join(__dirname, "tsconfig.esm.json"),
                 skipCodeGeneration: !aot,
