+++ tsconfig.json
@@ -10,6 +10,6 @@
     "files": [
         "plugins/PlatformFSPlugin.ts",
         "plugins/WatchStateLoggerPlugin.ts",
-        "host/platform.ts"
+        "host/resolver.ts"
     ]
 }
